refactor: SelectCategory 에서 카테고리만 선택 하도록 역할 분리 -  #13
