+++ package.json
@@ -16,6 +16,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.50.0",
+    "react-intersection-observer": "^9.8.2",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.21.3",
     "react-scripts": "5.0.1",
